pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "lightningviewer"
version = "0.7.3"
description = "CLI + API for downloading, querying and visualising Blitzortung lightning archives."
authors = [
  { name = "Sylvain Decarpentrie", email = "svdvet@orange.fr" }
]
readme = "README.md"
requires-python = ">=3.9"
license = "MIT" 

dependencies = [
  "python-dotenv",
  "requests",
  "tqdm",
  "pandas",
  "geopy",
  "streamlit",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[project.scripts]
lv = "lightningviewer.cli:main"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

